added PostDetail comment method and view

+++ posts/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.views.generic import ListView, DetailView, CreateView, UpdateView, DeleteView
 from .models import Post, PostView, Comment, Like
-from .forms import PostForm
+from .forms import PostForm, CommentForm
 
 
 class PostListView(ListView):
@@ -24,9 +24,27 @@ class PostCreateView(CreateView):
 class PostDetailView(DetailView):
     model = Post
 
+    def post(self, *args, **kwargs):
+        form = CommentForm(self.request.POST)
+        if form.is_valid():
+            post = self.get_object()
+            comment = form.instance
+            comment.user = self.request.user
+            comment.post = post
+            comment.save()
+            return redirect('detail', slug=post.slug)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context.update({
+            'form': CommentForm()
+        })
+        return context
+
     def get_object(self, **kwargs):
         object = super().get_object(**kwargs)
-        PostView.objects.get_or_create(user=self.request.user, post=object)
+        if self.request.user.is_authenticated:
+            PostView.objects.get_or_create(user=self.request.user, post=object)
         return object
 
 
@@ -49,14 +67,15 @@ class PostDeleteView(DeleteView):
 
 
 def like(request, slug):
-    post = get_object_or_404(Post, slug=slug)
-    like_qs = Like.objects.filter(user=request.user, post=post)
-    if like_qs.exists():
-        # unlike post
-        like_qs[0].delete()
-        return redirect('detail', slug=slug)
+    if request.user.is_authenticated:
+        post = get_object_or_404(Post, slug=slug)
+        like_qs = Like.objects.filter(user=request.user, post=post)
+        if like_qs.exists():
+            # unlike post
+            like_qs[0].delete()
+            return redirect('detail', slug=slug)
+        else:
+            new_like = Like.objects.create(user=request.user, post=post)
+            return redirect('detail', slug=slug)
     else:
-        new_like = Like.objects.create(user=request.user, post=post)
-        return redirect('detail', slug=slug)
-
-
+        return redirect('list')
